sofia/requirements.txt
torch>=2.2
transformers>=4.44
sentence-transformers>=3.0.0
datasets>=2.20
faiss-cpu>=1.7.4
rank_bm25>=0.2.2
huggingface_hub>=0.24.0
accelerate>=0.34.0
peft>=0.13.0
trl>=0.9.6
mteb>=1.15.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.8.0
scikit-learn>=1.4
numpy>=1.26
pdoc>=14.5.0
graphviz>=0.20.3
